pyproject.toml
[project]
name = "jira-readiness-agent"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "langgraph",
    "langchain",
    "langchain-anthropic",
    "anthropic",
    "langfuse",
    "httpx",
    "python-dotenv",
    "pyyaml",
    "jupyter",
]

[dependency-groups]
dev = ["ruff"]

[tool.ruff]
target-version = "py312"
line-length = 100
